Move 'Helpers' into sub component files
